refactor(gateway): split display-metadata patch mutations into their own module
The color branch pushed sessions-patch.ts over the max-lines budget; the
label/icon/color/category/boardFace block moves to
sessions-patch-display-metadata.ts unchanged. Restore the catalog streaming
test to main (color projection is covered at the entry-snapshot owner) and
satisfy the hovercard row's new required hasAutomation field in the catalog
hovercard test.
